Agree on byte ordering with world-id-contracts
This project was sending data encoded using little endian ordering where
Solidity expects big-endian. We decided to swap semaphore-mtb for two
reasons:

- It makes sense to use network byte ordering for a networked protocol.
- We want to avoid doing as much computation as possible on the EVM to
  reduce gas costs, so avoiding the need to twiddle bits in the contract
  is a win on that front as well.

The implementation ensures that the byte reordering is performed outside
of the arithmetic circuit, doing the remapping purely in Go. This means
that we do not incur any performance hit when executing the circuit.