prover: insertion_circuit: adjust web API to 4844
Model insertion circuit inputs calculation after the existing insertion
circuit tests. As a result, move InputHash from input to the response of
the service. The response is more sophisticated now and is modeled
accordingly, to satisfy contract's needs. Provide JSON (un)marshaling
methods for the new response type, so it can be returned from the
service.

Move some of the helper functions from insertion circuit test to circuit
utils, as they're now used in the proving system for 4844 calculations.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>