insertion_circuit: use versioned KZG hash
As per the EIP-4844 specs, KZG commitment should not be passed as raw
bytes. Instead, a versioned hash of the commitment should be used.
Implement the hashing function according to the specs and apply it on
the commitment. Use the hash instead of the raw commitment for
evaluation point calculation and pass the hash to the circuit instead.

Reduce the evaluation point value at the input to the circuiut as an
optimization, to avoid passing around large values.

Introduce relevant changes in the insertion circiut test as well as the
integration test and the prover itself. Export the BN254 reduction
function to the circuit utils file, as now it's used in the integration
test, which is outside of the prover package.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>