prover: verify: update `verify` command to EIP-4844
`gnark-mbu verify` was neglected in EIP-4844 changes. Update the
VerifyInsertion handler, so that it uses the recent insertion circuit
structure. Fields of the structure are populated both from:
 - standard input (previously only proof was provided via stdin, now
   it's full response, which also contains proof),
- new --params flag (previously parameters were not needed).

The new --params flag accepts input from `gnark-mbu gen-test-params`
command.

Dedicated --input-hash flag is not needed for the insertion mode, as
input hash is included in the response. This flag was made optional and
is checked in deletion only.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>